requirements.txt
# requirements.txt (consolidado para deploy no Render)
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.11
psycopg==3.2.10
# psycopg2-binary removed; using psycopg v3
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
gunicorn==21.2.0